cleaned up run_pyrate from the remaining unused functions
